requirements.txt
shapely==1.6.4.post2
fiona==1.8.6
pandas==0.24.2
geopandas==0.5.0
opencv-python==4.1.0.25
numpy==1.16.3
tqdm==4.32.1
GDAL==2.4.0  # pyup: ignore
rtree==0.8.3
networkx==2.3
rasterio==1.0.23
scipy==1.3.0
scikit-image==0.15.0
tensorflow==1.13.1
torch==1.1.0
affine==2.2.2
albumentations==0.2.3
rio-tiler==1.2.7
pyyaml==5.1
requests==2.22.0
